修改了数据的大小，由原来的int变成了现在的long long，能够承载更大的数据。另外去掉了atoi，atoll这些函数，改用stringstream统一做字符串和数字的转换。不过目前好像还只能够返回long类型的数据，似乎因为C中没有longlong

+++ Dll/src/calc.h
@@ -15,7 +15,7 @@ extern "C"
 #endif
 
     //对外调用的计算接口
-    DLL_EXPORT int calc(const char *s, int length);
+    DLL_EXPORT long calc(const char *s, unsigned int length);
 
 #ifdef RELEASE
 }
@@ -26,12 +26,12 @@ extern "C"
 #include "string"
 
 //从字符串转换为中缀表达式
-std::vector<std::string> str2In(const char *s, int length);
+std::vector<std::string> str2In(const char *s, unsigned int length);
 
 //从中缀表达式转换为后缀表达式
 std::vector<std::string> in2Post(const std::vector<std::string> &inFixExpression);
 
 //通过后缀表达式计算
-int calc(const std::vector<std::string> &postFixExpression);
+long long calc(const std::vector<std::string> &postFixExpression);
 
 #endif
